# expression: implement vectorized evaluation for 'builtinEQIntSig'