<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for builtinEQIntSig, for #12103


### What is changed and how it works?
```
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-VecBuiltinFunc-8         	  344095	      3118 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-NonVecBuiltinFunc-8      	   27732	     45131 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-VecBuiltinFunc#01-8      	  323122	      3431 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-NonVecBuiltinFunc#01-8   	   28254	     42421 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-VecBuiltinFunc#02-8      	  382246	      3217 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-NonVecBuiltinFunc#02-8   	   28332	     42380 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-VecBuiltinFunc#03-8      	  386941	      3043 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinEQIntSig-NonVecBuiltinFunc#03-8   	   28412	     42049 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test